Data structure player

+++ Assets/Script/PlayerMovement/HPBar.cs
@@ -183,7 +183,7 @@ public class HPBar : MonoBehaviour
 
     // ── Event Handler ─────────────────────────────────────────────────────────
 
-    void HandleHealthChanged(int currentHP)
+    void HandleHealthChanged(float currentHP)
     {
         if (playerHealth == null) return;
